refactor(frontend): extract theme toggle into a reusable component
