+++ apps/frontend/app/(dashboard)/layout.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { useTheme } from "next-themes";
 import type React from "react";
 import { useEffect } from "react";
 import { Sidebar } from "@/components/composites/sidebar/sidebar";
@@ -18,9 +17,6 @@ export default function DashboardLayout({
     const router = useRouter();
     const { logout } = useLogout();
 
-    const { theme, setTheme, resolvedTheme } = useTheme();
-    const isDark = (resolvedTheme ?? theme) === "dark";
-
     // Auth Protection
     useEffect(() => {
         if (!isPending && !session) {
@@ -28,20 +24,12 @@ export default function DashboardLayout({
         }
     }, [session, isPending, router]);
 
-    const toggleTheme = () => {
-        setTheme(isDark ? "light" : "dark");
-    };
-
     if (isPending) return <DashboardLoadingView />;
     if (!session) return null;
 
     return (
         <div className="flex h-screen w-screen overflow-hidden bg-slate-50 font-sans transition-colors duration-300 dark:bg-slate-950 dark:text-slate-200">
-            <Sidebar
-                onToggleTheme={toggleTheme}
-                onLogout={logout}
-                isDark={isDark}
-            />
+            <Sidebar onLogout={logout} />
 
             <main className="relative flex h-full w-full overflow-hidden">
                 {children}
